third_party/blink/web_tests/external/wpt/fullscreen/api/document-fullscreen-keyboard-lock-change.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] Requesting fullscreen again without keyboard lock should disable it
  assert_equals: fullscreen should deactivate expected null but got Element node <body><div id="log"></div>\n<script src="/fullscreen/api/d...
Harness: the test ran to completion.
